backend/requirements.txt
Flask
flask-cors
numpy
Pillow
matplotlib
scikit-image
scipy
PyWavelets
gunicorn

